requirements.txt
google-generativeai
PyMuPDF  # For parsing PDFs
python-docx  # For parsing DOCX
